+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Moved `gzip` package to a better suited package `io.gzip`.
 - Added `Hex.string(int, int)` to convert an integer to hexadecimal string with the given width.
+- Added `jmh-core` version `1.37` for benchmarking.
+- Added `jmh-generator-annprocess` version `1.37` for benchmarking.
 
 ---